feature: 原子操作能力支持滚动执行 #446

+++ src/backend/job-execute/service-job-execute/src/main/java/com/tencent/bk/job/execute/dao/impl/GseTaskDAOImpl.java
@@ -102,7 +102,7 @@ public class GseTaskDAOImpl implements GseTaskDAO {
          * if an existing row is updated, and 0 if an existing row is set to its current values.
          */
         if (result.size() > 0) {
-            id = result.stream().map(record -> record.get(TABLE.ID)).findFirst().get().intValue();
+            id = result.stream().map(record -> record.get(TABLE.ID)).findFirst().get();
         }
         return id;
     }
